update drupal/core to 10.3.13 and contrib theme and modules

+++ html/core/modules/text/src/Plugin/Field/FieldType/TextItemBase.php
@@ -97,7 +97,7 @@ abstract class TextItemBase extends FieldItemBase {
    * {@inheritdoc}
    */
   public function applyDefaultValue($notify = TRUE) {
-    // @todo: Add in the filter default format here.
+    // @todo Add in the filter default format here.
     $this->setValue(['format' => NULL], $notify);
     return $this;
   }
@@ -138,7 +138,7 @@ abstract class TextItemBase extends FieldItemBase {
     }
     else {
       // Textfield handling.
-      $max = ceil($settings['max_length'] / 3);
+      $max = (int) ceil($settings['max_length'] / 3);
       $value = substr($random->sentences(mt_rand(1, $max), FALSE), 0, $settings['max_length']);
     }
